Increased the possible graduations of a value/gradual representation (previously overflow could occur when having more than 255 steps).
git-svn-id: https://svn.52north.org/svn/ilwis/drawers@2476 b64c5c0c-1e74-440b-9268-4ad09cf26390

+++ Sources/Engine/Representation/RPRGRAD.cpp
@@ -221,7 +221,7 @@ Color RepresentationGradual::clrRaw(long iRaw) const
 	return col;  
 }
 
-byte RepresentationGradual::iColor(double rValue) const
+int RepresentationGradual::iColor(double rValue) const
 {
 	if (rValue < arLimits[0])
 		return 1;
